requirements-dev.txt
-r requirements.txt

black==23.1.0
isort==5.12.0
pytest==7.1.2
factory-boy==3.2.1
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-xdist==3.3.1
